requirements.txt
autopep8==2.0.0
click==8.1.3
colorama==0.4.6
DateTime==4.7
Flask==2.2.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
pycodestyle==2.10.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
SQLAlchemy-Utils==0.38.3
tomli==2.0.1
Werkzeug==2.2.2
zope.interface==5.5.2
